second update of plot code The second update is aimed at simpler plot configurations, better flexibility, faster plot generation and reduce memory usage in JDBC connection and plotting processing by aggravating large number of jdbc rows into fixed number of bins.
Major changes:

1. Changed DAO JDBC interface to use generics;
2. Simplified rendering/xml configuration, replaced AxisInfo and ColumnDescriptors by automatic or simpler configuration, more default values added;
3. Changed rendering methods by using aggregation to reduce memory usage;
4. Added statistics to plot (min, max, mean, std);
5. Added bin size parameter, StartDate and EndDate no longer optional (needed for histogram generation).
Timur commits on behalf of Tao
http://rb.dcache.org/r/2275
Approved by : timur

+++ modules/common/org/dcache/commons/plot/renderer/svg/SVGGradient.java
@@ -41,8 +41,8 @@ public class SVGGradient implements SVGColor{
     * @param color2 color of end offset
     * @param opacity2 opacity of end offset
     */
-    public void setLinerGradient(int x1, int y1, RGBColor color1, float opacity1,
-            int x2, int y2, RGBColor color2, float opacity2){
+    public void setLinerGradient(float x1, float y1, RGBColor color1, float opacity1,
+            float x2, float y2, RGBColor color2, float opacity2){
        element.setAttribute("x1", x1 + "%");
        element.setAttribute("y1", y1 + "%");
        element.setAttribute("x2", x2 + "%");
